requirements.txt
appdirs==1.4.3
asgiref==3.3.4
beautifulsoup4==4.9.3
CacheControl==0.12.6
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cssselect2==0.4.1
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.2
django-admin-interface==0.16.3
django-bootstrap4==3.0.0
django-ckeditor==6.1.0
django-colorfield==0.4.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.2.2
django-object-actions==3.0.2
html5lib==1.0.1
idna==2.8
install==1.3.4
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.2.0
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
pyparsing==2.4.6
pyphen==0.11.0
python-decouple==3.4
pytoml==0.1.21
pytz==2021.1
reportlab==3.5.67
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
Unidecode==1.2.0
urllib3==1.25.8
WeasyPrint==52.5
webencodings==0.5.1
